+++ app/redux/store.js
@@ -9,24 +9,29 @@ import connection from './connection/reducers';
 import connectionActions from './connection/actions';
 import settings from './settings/reducers';
 import settingsActions from './settings/actions';
+import support from './support/reducers';
+import supportActions from './support/actions';
 
 import type { Store } from 'redux';
 import type { History } from 'history';
 import type { AccountReduxState } from './account/reducers';
 import type { ConnectionReduxState } from './connection/reducers';
 import type { SettingsReduxState } from './settings/reducers';
+import type { SupportReduxState } from './support/reducers';
 
-import type { ConnectionAction } from './connection/actions';
 import type { AccountAction } from './account/actions';
+import type { ConnectionAction } from './connection/actions';
 import type { SettingsAction } from './settings/actions';
+import type { SupportAction } from './support/actions';
 
 export type ReduxState = {
   account: AccountReduxState,
   connection: ConnectionReduxState,
   settings: SettingsReduxState,
+  support: SupportReduxState,
 };
 
-export type ReduxAction = AccountAction | SettingsAction | ConnectionAction;
+export type ReduxAction = AccountAction | ConnectionAction | SettingsAction | SupportAction;
 export type ReduxStore = Store<ReduxState, ReduxAction, ReduxDispatch>;
 export type ReduxGetState = () => ReduxState;
 export type ReduxDispatch = (action: ReduxAction | ReduxThunk) => any;
@@ -42,6 +47,7 @@ export default function configureStore(
     ...accountActions,
     ...connectionActions,
     ...settingsActions,
+    ...supportActions,
     pushRoute: (route) => push(route),
     replaceRoute: (route) => replace(route),
   };
@@ -50,6 +56,7 @@ export default function configureStore(
     account,
     connection,
     settings,
+    support,
     router: routerReducer,
   };